Alteração nos arquivos dividindo por processos
