+++ functions.c
@@ -14,7 +14,7 @@ Value listar_cliente(Conta contas[], int *pos, int *user) {
      printf("Lista");
 } // Função de Listar os clientes
 Value debito(Conta contas[], int *pos, int *user) {
-     printf("débito");
+     printf("debito");
 }         // Função de debitar dinheiro de uma conta
 Value deposito(Conta contas[], int *pos, int *user) {
      printf("deposito");
@@ -23,123 +23,8 @@ Value extrato(Conta contas[], int *pos, int *user) {
      printf("extrado");
 }       // Função de gerar extrato do cliente
 Value transacao(Conta contas[], int *pos, int *user) {
-     printf("transação");
+     printf("transacao");
 }     // Função de realizar transacao entre contas
-
-
-// Funções de Autenticação -----------------------------------------
-Value login(Conta contas[], int *pos, int *user){
-    printf("Login");
-    return OK;
-}; 
-Value cadastro(Conta contas[], int *pos, int *user){
-    printf("cadastro");
-    return OK;
-}; 
-
-Value hash(char senha[], int *pos){
-     printf("hash");
-}
-Value hash_compare(char hash[], char senha[], int *pos){
-     printf("compare");
-}
-
-// Menu
-int menu_login(){
-    int opcao;
-    
-    do{
-        printf("\n\033[34m");
-        printf("|==================================|\n");
-        printf("|- - -| Banco Quem Poupa Tem |- - -|\n");
-        printf("|==================================|\n");
-        printf("| > Menu de Login:                 |\n");
-        printf("|   1 - Fazer Login                |\n");
-        printf("|   2 - Fazer Cadastro             |\n");
-        printf("|   0 - Sair                       |\n");
-        printf("|==================================|\n");
-        printf("| > Escolha uma opcao: ");
-        scanf("%d", &opcao);
-        clearBuffer();
-        
-        if(opcao <= 2 && opcao >= 1){
-            return opcao;
-        }
-        else if(opcao == 0){
-            break;
-        }
-        else{
-            printf("\033[31m| > Opção Inválida, Tente novamente...\n");
-        }
-    }while(opcao != 0);
-    return 0;
-}
-
-int menu_adm(){
-    int opcao;
-    int validation = 0;
-    
-    do{
-        printf("\n\033[34m");
-        printf("|==================================|\n");
-        printf("|- - -| Banco Quem Poupa Tem |- - -|\n");
-        printf("|==================================|\n");
-        printf("| > Menu de Administrador:         |\n");
-        printf("|   1 - Depósito                   |\n");
-        printf("|   2 - Débito                     |\n");
-        printf("|   3 - Transferência              |\n");
-        printf("|   4 - Extrato                    |\n");
-        printf("|   5 - Criar Usuário              |\n");
-        printf("|   6 - Listar Usuários            |\n");
-        printf("|   7 - Deletar Usuário            |\n");
-        printf("|   0 - Sair                       |\n");
-        printf("|==================================|\n");
-        printf("| > Escolha uma opcao: ");
-        scanf("%d", &opcao);
-        clearBuffer();
-
-        if(opcao <= 7 && opcao >= 0){
-            validation = 1;
-        }
-        else{
-            printf("\033[31m| | > Opção Inválida, tente novamente.\n");
-        }
-    }while(validation = 0);
-    
-    return opcao;
-}
-
-int menu_user(){
-    int opcao;
-    int validation = 0;
-    
-    do{
-        printf("\n\033[34m");
-        printf("|==================================|\n");
-        printf("|- - -| Banco Quem Poupa Tem |- - -|\n");
-        printf("|==================================|\n");
-        printf("| > Menu de Usuário:               |\n");
-        printf("|   1 - Depósito                   |\n");
-        printf("|   2 - Débito                     |\n");
-        printf("|   3 - Transferência              |\n");
-        printf("|   4 - Extrato                    |\n");
-        printf("|   0 - Sair                       |\n");
-        printf("|==================================|\n");
-        printf("| > Escolha uma opcao: ");
-        scanf("%d", &opcao);
-        clearBuffer();
-
-        if(opcao <= 4 && opcao >= 0){
-            validation = 1;
-        }
-        else{
-            printf("\033[31m| | > Opção Inválida, tente novamente.\n");
-        }
-    }while(validation = 0);
-    
-    return opcao;
-}
-
 
 //Funções auxiliares
 void tratarRes(Value err) {}; // Tratamento de Erros das funções de tipo Value
